+++ src/template/template.ts
@@ -1,5 +1,6 @@
 import { TemplateFile } from './template-file';
 import { TemplateParameter } from './template-parameter';
+import { TemplateCustomVariable } from './template-custom-variable';
 
 /**
  * Interface for a template group
@@ -24,7 +25,16 @@ export interface Template
 	/** Parameters that can be provided to the template */
 	parameters?: TemplateParameter[];
 
-	/** Directory where the template is located */
+	/** Inline variable definitions */
+	variables?: TemplateCustomVariable[];
+
+	/** Path to the external script relative to the repository root. */
+	variablesScript?: string;
+
+	/** Function name to call (default: "generateVariables") */
+	variablesFunction?: string;
+
+	/** Directory where the template is located (added at runtime) */
 	directory?: string;
 
 	/** Repository ID where this template is from (added at runtime) */
@@ -32,4 +42,7 @@ export interface Template
 
 	/** Repository display name (added at runtime) */
 	repositoryName?: string;
+
+	/** The compiled variables function (added at runtime) */
+	variablesFn?: Function;
 }