+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "codeTemplateHub" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## 0.2.0
 
-- Initial release
+### New Features
+
+- **Custom Variable Support**: Added functionality to define and use custom variables within templates. The variables can be define in the templates.json or by a custom javascript file.
+
+### Other Changes
+
+- Updated README with custom variable documentation